refactor: reorganize project structure for Mocha and Playwright compatibility
- Migrate app code and tests into separate directories for better organization.
- Update `package.json` files to adjust dependencies and scripts for Mocha and Playwright.
- Refactor `.gitignore` files to reflect new directory structure.
- Clean up and reformat TypeScript files for consistent indentation and readability.
