+++ TypeScript-Mocha/.mocharc.json
@@ -1,0 +1,4 @@
+{
+  "require": ["ts-node/register"],
+  "spec": "**/*.spec.ts"
+}
