+++ TypeScript-Mocha/tsconfig.json
@@ -13,5 +13,7 @@
     "downlevelIteration": true,
     "experimentalDecorators": true
   },
-  "exclude": ["node_modules"]
+  "exclude": [
+    "node_modules"
+  ]
 }
