--- TypeScript/app/index.ts
@@ -1,1 +1,0 @@
-export { Battery, ElectricRose } from "./electric-rose";
