--- TypeScript/app/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "@electric-rose/app",
-  "version": "1.0.0",
-  "description": "Electric Rose Battery Tracker - Core Application",
-  "main": "index.ts",
-  "types": "index.ts",
-  "scripts": {
-    "build": "tsc",
-    "start": "ts-node app.ts"
-  },
-  "author": "Peter 'Code Cop' Kofler",
-  "license": "BSD-3-Clause",
-  "devDependencies": {
-    "@types/node": "~22.15.21",
-    "ts-node": "~10.9.2",
-    "typescript": "~5.8.3"
-  }
-}
