--- TypeScript/test/mocha/mocha.opts
@@ -1,4 +1,0 @@
---compilers ts-node/register
---require source-map-support/register
---recursive
-test/**/*.spec.ts
