--- TypeScript/test/mocha/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@electric-rose/mocha-tests",
-  "version": "1.0.0",
-  "description": "Electric Rose Battery Tracker - Mocha Tests",
-  "author": "Peter 'Code Cop' Kofler",
-  "license": "BSD-3-Clause",
-  "scripts": {
-    "test": "mocha -r ts-node/register '**/*.spec.ts'"
-  },
-  "dependencies": {
-    "@electric-rose/app": "file:../../app"
-  },
-  "devDependencies": {
-    "@types/chai": "5.2.2",
-    "@types/mocha": "10.0.10",
-    "@types/node": "~22.15.21",
-    "chai": "~5.2",
-    "mocha": "~11.4",
-    "source-map-support": "0.5.21",
-    "ts-node": "~10.9.2",
-    "typescript": "~5.8.3"
-  }
-}
